Funcionalidad Cliente

+++ frontend/src/app/components/clientes/clientes.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Icliente } from "../../interfaces/icliente";
 import { ClientesService } from "../../services/clientes.service";
 import { Router } from "@angular/router";
+import Swal from 'sweetalert2';
+
 
 
 @Component({
@@ -11,20 +13,20 @@ import { Router } from "@angular/router";
 })
 export class ClientesComponent implements OnInit {
 
-  clientes : Icliente[];
-  flagClienteNuevo : boolean = false;
+  clientes: Icliente[];
+  flagClienteNuevo: boolean = false;
   message: string;
   flagClienteEliminado = false;
   flagProblemasTecnicos = false;
 
-  constructor( private _clientesService: ClientesService, private _router: Router) { }
+  constructor(private _clientesService: ClientesService, private _router: Router) { }
 
   ngOnInit() {
 
     //this.clientes = this._clientesService.getClientes().subscribe();
     this.getClientes();
-    
-    
+
+
     /*this._clientesService.getClientesPathVariable('chikitina').subscribe(
       response=>{console.log(response)},
       error=>{this.handleResponseError(error)});*/
@@ -32,63 +34,82 @@ export class ClientesComponent implements OnInit {
   }
 
 
-  handleResponseError(error){
+  handleResponseError(error) {
     console.log(error);
     console.log(error.error);
     console.log(error.message);
-    
+
 
 
   }
 
-  getClientes(){
+  getClientes() {
     this._clientesService.getClientes().subscribe(
-      response =>{
+      response => {
         (console.log(response));
         this.clientes = response;
       },
-      error=>{
+      error => {
         console.log('Error al consultar clientes');
         this.flagProblemasTecnicos = true;
         this.message = 'Problemas tecnicos presentados. Por favor contacte a soporte tecnico.';
 
-    }
+      }
     );
 
 
   }
 
 
-  getCliente(idCliente: string){
+  getCliente(idCliente: string) {
     console.log(idCliente);
-    this._router.navigate(['cliente',idCliente]);
+    this._router.navigate(['cliente', idCliente]);
 
   }
 
-  addCliente(){
+  addCliente() {
     console.log('adicionar cliente');
     this.flagClienteNuevo = true;
-
     this._router.navigate(['cliente']);
   }
 
-  deleteCliente(idCliente : string){
-    console.log("Eliminar cliente:" +idCliente);
-    this._clientesService.deleteCliente(idCliente).subscribe(
-      response =>{
-        console.log(response);
-        this.message = `Cliente con cedula ${idCliente} eliminado exitosamente`;
-        this.flagClienteEliminado = true;
-        this.getClientes();
-      },
-      error =>{this.handleResponseError;}
-    );
+  deleteCliente(idCliente: string) {
+    console.log("Eliminar cliente:" + idCliente);
+    Swal.fire({
+      title: 'Está seguro?',
+      text: "Usted no podra revertir esta operacion!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'No, cancelar!',
+      
+      confirmButtonText: 'Si, eliminarlo!'
+    }).then((result) => {
+      if (result.value) {
+        this._clientesService.deleteCliente(idCliente).subscribe(
+          response => {
+            //this.message = `Cliente con cedula ${idCliente} eliminado exitosamente`;
+            this.flagClienteEliminado = true;
+            Swal.fire(
+              'Eliminar cliente',
+              `Cliente con CC ${idCliente} eliminado exitosamente!`,
+              'success'
+            )
+            this.getClientes();
+
+          },
+          error => { this.handleResponseError; }
+        );
+
+      }
+    })
 
   }
 
-  updateCliente(idCliente: string){
+  updateCliente(idCliente: string) {
     console.log(`Actualizar cliente No. ${idCliente}`);
-    this._router.navigate(['cliente',idCliente]);
+    this._router.navigate(['cliente', idCliente]);
 
 
   }
